backend/requirements.txt
fastapi>=0.100.0
elevenlabs>=0.2.27
websockets>=11.0
openai>=1.0.0
python-dotenv>=1.0.0
uvicorn>=0.23.0
speechbrain>=0.5.15
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
soundfile>=0.12.1
transformers>=4.30.0
aiohttp
httpx>=0.24.0
montreal-forced-aligner==2.0.6
